add auth's constraint

+++ internal/auth/interop/base.go
@@ -22,6 +22,10 @@ func (a AuthInterop) Create(ctx context.Context, token string, aut *auth.Auth) e
 }
 
 func (a AuthInterop) GetById(ctx context.Context, token string, id string) (*auth.Auth, error) {
+	_, err := a.ucase.Verify(ctx, token)
+	if err != nil {
+		return nil, err
+	}
 	return a.ucase.GetById(ctx, id)
 }
 
@@ -65,6 +69,17 @@ func (a AuthInterop) ChangeRole(ctx context.Context, token string, roleId string
 }
 
 func (a AuthInterop) Delete(ctx context.Context, token string, id string) error {
+	record, err := a.ucase.Verify(ctx, token)
+	if err != nil {
+		return err
+	}
+	authData, err := a.ucase.GetById(ctx, record.UID)
+	if err != nil {
+		return err
+	}
+	if authData.RoleId != auth.RoleAdmin {
+		return auth.ErrAuthNotAuthorized
+	}
 	return a.ucase.Delete(ctx, id)
 }
